Clean up and reformat
