+++ node/src/main.rs
@@ -8,7 +8,6 @@
 mod benchmarking;
 mod chain_spec;
 mod cli;
-mod client;
 mod command;
 mod eth;
 mod rpc;
